Score each guess.
The guesses are scored based on how restrictive they are.
The more possible solutions it eliminates the better a guess is.
This is just some rudimentary way of doing this to start off. There are
lots of ways we can score here.

There are some caveats with the way I'm determining if 2 words have the
same score. It might not be 100% correct but it is close enough for now.
If a letter appears twice in a word the output is deterministically but
arbitrary chosen from multiple possible representations. Those may not
be stable between words (but should be). If they are not stable there is
a bug here where we are overcounting the score. To fix: update to a
better match representation that uses the character as the key rather
than the index of the word as the key.