Fix the tests!
Yikes. This is still kind of a mess.
Added a data file for the unit tests so we don't mess with those.
I run the tests by cd ing into the src folder. Somethign is up with the relative imports that I  don't understand.
For some reason the linter keeps rying to add gibberish to the ends of the test files. There is some type of bug there.
The wordle tests inits the profiler lol which it shouldn't do. I'll have to update the design there...