Caching corpus maps. Using set intersections.
We can't cache the corpus reductions since we are using
differently populated corpuses. But instead I just use the same corpus and then get the intersection of the successive opersations.  This allows for caching, but we do have to
got through slightly larger corpuses. Seems to be a bit faster.
But now all our time is spent in set operations.
Since our corpus for a given run is fixed we can avoid the set operations.
Just represent it all as a 1-hot vector... That's next.