Improve performance of bit counter.
The bit counter now uses a bitmask and extract procedure to chunk
large binary words into smaller pieces.  The smaller chunks are
wrapped in a cache which will now have a near 100% hit rate.
This brings our bit counting to O(1) time given a fixed bitmap of Corpus size. The lookups are constant time and the splitting is O(log) itme but we have a fied dictioanry size so the splitting is a constant overhead. yay. This was fun.